+++ API/src/StellarPath.API/StellarPath.API.Core/Models/Destination.cs
@@ -5,7 +5,7 @@
         public int DestinationId { get; set; }
         public required string Name { get; set; }
         public int SystemId { get; set; }
-        public required int DistanceFromEarth { get; set; }
+        public required long DistanceFromEarth { get; set; }
         public required bool IsActive { get; set; }
 
     }
